Upgrade spark versions in test suite (#42)
* update spark versions

* add py38

* update pyarrow version

* pyarrow to 0.16.0

* only run spark 2.4 test for python < 3.8

* add test skip log

* update version compatibility in README

+++ build_tools/spark_2_4.sh
@@ -1,13 +1,21 @@
-# install spark
-PATH=$(echo "$PATH" | sed -e 's/:\/usr\/local\/lib\/jvm\/openjdk11\/bin//')
-JAVA_HOME=/usr/lib/jvm/java-1.8.0-openjdk-amd64
-mkdir -p /opt
-wget -q -O /opt/spark.tgz http://www.gtlib.gatech.edu/pub/apache/spark/spark-2.4.5/spark-2.4.5-bin-hadoop2.7.tgz
-tar xzf /opt/spark.tgz -C /opt/
-rm /opt/spark.tgz
-export SPARK_HOME=/opt/spark-2.4.5-bin-hadoop2.7
-export PATH=$PATH:/opt/spark-2.4.5-bin-hadoop2.7/bin
+python_version=$(python -c 'import sys; print(sys.version_info[:][1] < 8)')
 
-# run tests
-pip install -e .[tests]
-build_tools/test_script.sh
+# only run for python version under 3.8
+if [ $python_version == "True" ]
+then
+  # install spark
+  PATH=$(echo "$PATH" | sed -e 's/:\/usr\/local\/lib\/jvm\/openjdk11\/bin//')
+  JAVA_HOME=/usr/lib/jvm/java-1.8.0-openjdk-amd64
+  mkdir -p /opt
+  wget -q -O /opt/spark.tgz http://www.gtlib.gatech.edu/pub/apache/spark/spark-2.4.6/spark-2.4.6-bin-hadoop2.7.tgz
+  tar xzf /opt/spark.tgz -C /opt/
+  rm /opt/spark.tgz
+  export SPARK_HOME=/opt/spark-2.4.6-bin-hadoop2.7
+  export PATH=$PATH:/opt/spark-2.4.6-bin-hadoop2.7/bin
+
+  # run tests
+  pip install -e .[tests]
+  build_tools/test_script.sh
+else
+  echo "Skipped tests - spark and python versions are not compatible"
+fi
